Remove AV, Numpy, Pillow Code + onefile.py
Remove all dependencies from AV, Numpy, and Pillow to make the library
work with lighter devices.

In addition, add `onefile.py` which is a 100% working example of simply
taking off and landing by putting all library code in one file. This was
done because some of the bad python emulators on android don't even
allow you to import files locally.